requirements.txt
# === Core LLM / Agent Framework ===
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-ollama>=0.1.0
requests>=2.31.0

# === Hindi / Indic NLP ===
# NOTE: indic-nlp-library resources are NOT on PyPI.
#       Run: python setup_resources.py  after installing requirements.
indic-nlp-library>=0.91

# === NLP / Transformers ===
transformers>=4.40.0
sentence-transformers>=2.7.0
datasets>=2.19.0
torch>=2.2.0
accelerate>=0.29.0
evaluate>=0.4.1
seqeval>=1.2.2

# === Vector Store ===
faiss-cpu>=1.8.0

# === Web Scraping ===
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=5.2.0

# === Wikipedia XML Parsing ===
mwparserfromhell>=0.6.6
bz2file>=0.98

# === API Layer ===
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# === UI ===
streamlit>=1.35.0

# === Visualization ===
plotly>=5.22.0

# === Word Cloud ===
wordcloud>=1.9.3
Pillow>=10.3.0
matplotlib>=3.9.0

# === Data Processing ===
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.13.0
tqdm>=4.66.0

# === Utilities ===
python-dotenv>=1.0.0
pydantic>=2.7.0
click>=8.1.0
rich>=13.7.0

# === Testing ===
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
